# perf: Add criterion benchmark for aggregate expressions